+++ request.rest
@@ -43,7 +43,7 @@ Content-Type: application/json
     "role":"ADMIN"
 }
 
-# ================================================ Making Login ================================================ #
+# ================================================ MAKING LOGIN ================================================ #
 
 ###
 POST http://localhost:3003/users/login
@@ -63,7 +63,7 @@ Content-Type: application/json
     "password": "87654321"
 }
 
-# ================================================ Get Profile ================================================ #
+# ================================================ GET PROFILE ================================================ #
 
 ### 
 GET http://localhost:3003/users/profile
