+++ tests/mocks/services/AuthenticatorMock.ts
@@ -1,5 +1,5 @@
-import { IAuthenticator } from "../../src/business/ports";
-import { UserRole } from "../../src/model/User";
+import { IAuthenticator } from "../../../src/business/ports";
+import { UserRole } from "../../../src/model/User";
 
 export class AuthenticatorMock implements IAuthenticator {
     public generateToken = jest.fn( () => {
