+++ tests/mocks/services/IdGeneratorMock.ts
@@ -1,4 +1,4 @@
-import { IIdGenerator } from "../../src/business/ports";
+import { IIdGenerator } from "../../../src/business/ports";
 
 export class IdGeneratorMock implements IIdGenerator {
     public generate = jest.fn( () => {
